moved the camera loop from the profiler preparation script into the gui
